requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.3
django-cors-headers==3.13.0
django-countries==7.5.1
django-environ==0.9.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
phonenumberslite==8.13.6
Pillow==9.3.0
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.6
requests==2.28.2
sqlparse==0.4.3
stripe==5.2.0
typing_extensions==4.5.0
tzdata==2022.6
urllib3==1.26.14
